+++ file_mapper/Makefile
@@ -1,0 +1,7 @@
+.PHONY: test watch
+
+test:
+	pytest -v
+
+watch:
+	find . -type f -name "*.py" | entr -p python file_mapper.py /_
